feat(typeahead): fix tabbing through typeahead component
The typeahead component clears the selected item when it receives focus. Added a configuration to
all it to not clear the selected item when tabbed into but it will still clear it and show results
if the value in the input is empty string or if it is clicked. Made the new default to not clear
item.